backend/trainer-portal-service/requirements.txt
# ============================================================
# trainer-portal-service/requirements.txt
# Production-grade Python FastAPI backend
# ============================================================

# Web framework
fastapi==0.111.1
uvicorn[standard]==0.30.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2          # Optional: if not using Liquibase for Python models

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Validation
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0

# Env & config
python-dotenv==1.0.1

# Middleware & security
python-multipart==0.0.9

# Logging
loguru==0.7.2

# HTTP client (for health checks in tests)
httpx==0.27.0

# ─── Dev / Test ────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
anyio==4.4.0
